refactor(agentd): extract provider-usage parsing into internal package
cmd/agentd/main.go was 5708 lines across 160 functions. Move the provider usage
CLI parsing -- `claude usage`, `codex status`, `gemini stats` output and its
supporting number/time helpers -- into internal/providerusage, bringing main.go
down to 4813 lines.

This is deliberately distinct from internal/usage, which parses usage out of
tmux console snapshots. Same domain, different input: the commands here emit
JSON or a stable report layout so the parsers can be strict, while the snapshot
parsers must tolerate arbitrary terminal text. The package doc records the
distinction so the two are not mistaken for duplicates.

main_usage_parse_test.go moves with the code it covers. Remaining seams for a
later pass: CLI commands, the ~25 executeXxx tmux methods, and MCP handlers.
